[Parser] Be far more precise when matching HTML
I found that after trying to use this utility to highlight it's own
repo, some lines were being messed up because of crap where a line would
check if a substring was in the current line, but the current line was
the line in the code that checks if that substring is in the line, etc.
By using a pretty strict regex you can deal with this issue, but in the
future this program should probably use a proper XML/HTML parser.